Health fully split and full game builds
