+++ src/lib-tempo/src/system/SystemHealth.cpp
@@ -1,4 +1,4 @@
-#include <tempo/entity/SystemHealth.hpp>
+#include <tempo/system/SystemHealth.hpp>
 
 namespace tempo {
